Refactor : participation 디렉토리 game 디렉토리로 병합
