+++ src/main/java/sportsmatchingservice/game/repository/ParticipationRepository.java
@@ -1,7 +1,7 @@
 package sportsmatchingservice.game.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sportsmatchingservice.participation.domain.Participation;
+import sportsmatchingservice.game.domain.Participation;
 
 public interface ParticipationRepository extends JpaRepository<Participation, Long> {
